Adds:
1. avg ms/op (in addiiton to existing percentiles)
2. percentiles for MB/s (before just median)
3. avg for MB/s
4. clearer labeling/spacing

Sample output line
```
populate 1024x1000 (clean, indexes: 0) 50/75/90/95%=146.80,151.20,178.50,178.50 ms/op avg=153.42 ms/op 50/75/90/95%=6.65,6.46,5.47,5.47 MB/s avg=6.37 MB/s (5 runs sampled)
```