@aboodman 2021-10-29
Nice but is there a reason to include the ops/sec? Is anyone (@arv?) advocating for it?