@grgbkr 2021-11-02
I've rolled the parts worth keeping of this into https://github.com/rocicorp/replicache/pull/654, so going to abandon this pull request.